+++ src/app.ts
@@ -16,6 +16,38 @@ import cors from 'cors';
 const app: express.Application = express();
 const allowedOrigins = ['http://localhost:3000', 'http://10.100.102.172:3000', 'http://10.100.102.172:5000'];
 
+
+
+//all what I add
+//////////////////////////////////////////////////////
+// Validate FIREBASE_PRIVATE_KEY
+import admin from "firebase-admin";
+
+// בדוק אם Firebase Admin מאותחל כבר
+if (!admin.apps.length) {
+  const serviceAccount = {
+    type: process.env.FIREBASE_TYPE || "service_account",
+    project_id: process.env.FIREBASE_PROJECT_ID || "",
+    private_key_id: process.env.FIREBASE_PRIVATE_KEY_ID || "",
+    private_key: (process.env.FIREBASE_PRIVATE_KEY || "").replace(/\\n/g, '\n'),
+    client_email: process.env.FIREBASE_CLIENT_EMAIL || "",
+    client_id: process.env.FIREBASE_CLIENT_ID || "",
+    auth_uri: process.env.FIREBASE_AUTH_URI || "",
+    token_uri: process.env.FIREBASE_TOKEN_URI || "",
+    auth_provider_x509_cert_url: process.env.FIREBASE_AUTH_PROVIDER_CERT_URL || "",
+    client_x509_cert_url: process.env.FIREBASE_CLIENT_CERT_URL || "",
+  };
+
+  admin.initializeApp({
+    credential: admin.credential.cert(serviceAccount as admin.ServiceAccount),
+  });
+}
+////////////////////////////////////////////////////////////////////////
+
+
+
+
+
 const mongoURI: string = process.env.MONGO_URI_STAGE || 'mongodb://localhost:27017/Hikemeet';
 mongoose
   .connect(mongoURI)
@@ -80,8 +112,8 @@ mongoose
       console.info(`Server is listening on ${bind}`);
     }
 
-    app.listen(3000, '0.0.0.0', () => {
-      console.log('Server is running on http://0.0.0.0:3000');
+    app.listen(5000, '0.0.0.0', () => {
+      console.log('Server is running on http://0.0.0.0:5000');
     });
     server.on('error', onError);
     server.on('listening', onListening);
